# docs/API-SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: PSTN2 API
  description: |
    Distributed telecommunications protocol API for fraud prevention,
    direct routing, and enhanced emergency services.

    **Features:**
    - Real-time caller ID authentication
    - Direct peer-to-peer routing
    - End-to-end encryption
    - Emergency location services
    - Distributed directory

    **Base URL:** `https://api.{your-cp}.com/pstn2/v1`
  version: 1.0.0
  contact:
    name: PSTN2 Support
    url: https://pstn2.org
    email: nick.holland@8x8.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.cp1.example.com/pstn2/v1
    description: Example CP1 Server
  - url: https://api.cp2.example.com/pstn2/v1
    description: Example CP2 Server
...
paths:
  /auth/verify:
    post:
      tags:
        - Authentication
      summary: Verify caller ID (Direct Query)
      description: |
        Verify that a call originated from the claiming CP.
        Used by recipient CP to authenticate inbound calls in real-time.

        **Performance Target:** < 100ms response time
      operationId: verifyCall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CallVerificationRequest'
            examples:
              basic:
                $ref: '#/components/examples/VerificationRequestBasic'
      responses:
        '200':
          description: Verification successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallVerificationResponse'
              examples:
                verified:
                  $ref: '#/components/examples/VerificationResponseVerified'
        '301':
          description: Number ported to different CP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortingResponse'
        '404':
          description: Call not found (potential fraud)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  $ref: '#/components/examples/CallNotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/tokens:
    post:
      tags:
        - Authentication
      summary: Create authentication token (Token Pool)
      description: |
        Create a short-lived token in the shared token pool.
        Used before placing an outbound call.

        **TTL:** 30-60 seconds typical
      operationId: createToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenCreateRequest'
      responses:
        '201':
          description: Token created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenCreateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/tokens/{tokenId}:
    get:
      tags:
        - Authentication
      summary: Verify authentication token
      description: |
        Retrieve and verify a token from the shared pool.
        Used by recipient CP on inbound call.
      operationId: verifyToken
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
            pattern: '^TK-[A-Za-z0-9]{16}$'
          example: TK-abc123XYZ789def
      responses:
        '200':
          description: Token valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenVerifyResponse'
        '404':
          description: Token not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /routing/request:
    post:
      tags:
        - Routing
      summary: Request direct routing
      description: |
        Request connection details for direct peer-to-peer media.
        Includes capability negotiation and key exchange.

        **Performance Target:** < 200ms response time
      operationId: requestRouting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoutingRequest'
            examples:
              audio:
                $ref: '#/components/examples/RoutingRequestAudio'
              video:
                $ref: '#/components/examples/RoutingRequestVideo'
      responses:
        '200':
          description: Routing accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingResponse'
              examples:
                accepted:
                  $ref: '#/components/examples/RoutingResponseAccepted'
        '503':
          description: Service unavailable (capacity/maintenance)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingRejection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Destination number not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /directory/lookup/{phoneNumber}:
    get:
      tags:
        - Directory
      summary: Lookup number range holder
      description: |
        Find which CP currently holds a specific number.

        **Performance:** < 50ms (cache hit), < 200ms (cache miss)
      operationId: lookupNumber
      parameters:
        - name: phoneNumber
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/PhoneNumber'
          example: '+441234567890'
      responses:
        '200':
          description: CP found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryLookupResponse'
        '404':
          description: Number not found in directory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    cpSignature:
      type: apiKey
      in: header
      name: X-PSTN2-Signature
      description: Ed25519 signature of request body

    psapCertificate:
      type: mutualTLS
      description: PSAP client certificate authentication

    tokenPoolJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT for token pool operations
  schemas:
    PhoneNumber:
      type: string
      pattern: '^\+[1-9]\d{1,14}$'
      description: E.164 formatted phone number
      example: '+441234567890'
...
    Timestamp:
      type: string
      format: date-time
      description: ISO 8601 timestamp with timezone
      example: '2025-11-30T21:30:00.000Z'
...
    CallVerificationRequest:
      type: object
      required:
        - messageId
        - timestamp
        - version
        - requestingCP
        - callerID
        - calledID
        - callReference
      properties:
        messageId:
          type: string
          format: uuid
        timestamp:
          $ref: '#/components/schemas/Timestamp'
        version:
          type: string
          enum: ['1.0']
        requestingCP:
          $ref: '#/components/schemas/RCPID'
        callerID:
          $ref: '#/components/schemas/PhoneNumber'
        calledID:
          $ref: '#/components/schemas/PhoneNumber'
        callReference:
          $ref: '#/components/schemas/CallReference'
        signature:
          $ref: '#/components/schemas/Base64String'
    CallVerificationResponse:
      type: object
      required:
        - verified
        - callReference
        - timestamp
      properties:
        verified:
          type: boolean
        callReference:
          $ref: '#/components/schemas/CallReference'
        callerName:
          type: string
          example: 'John Smith'
        callerOrg:
          type: string
          example: 'ACME Corp'
        callPurpose:
          type: string
          example: 'Account verification'
        trustLevel:
          type: string
          enum: [low, medium, high, verified]
        branding:
          $ref: '#/components/schemas/BrandingInfo'
        timestamp:
          $ref: '#/components/schemas/Timestamp'
        signature:
          $ref: '#/components/schemas/Base64String'
...
    PortingResponse:
      type: object
      required:
        - verified
        - ported
        - currentHolder
        - callReference
      properties:
        verified:
          type: boolean
          enum: [false]
        ported:
          type: boolean
          enum: [true]
        currentHolder:
          $ref: '#/components/schemas/RCPID'
        portedAt:
          $ref: '#/components/schemas/Timestamp'
        callReference:
          $ref: '#/components/schemas/CallReference'
    TokenCreateRequest:
      type: object
      required:
        - messageId
        - timestamp
        - originatingCP
        - callerID
        - calledID
        - callReference
        - ttl
      properties:
        messageId:
          type: string
          format: uuid
        timestamp:
          $ref: '#/components/schemas/Timestamp'
        originatingCP:
          $ref: '#/components/schemas/RCPID'
        callerID:
          $ref: '#/components/schemas/PhoneNumber'
        calledID:
          $ref: '#/components/schemas/PhoneNumber'
        callReference:
          $ref: '#/components/schemas/CallReference'
        ttl:
          type: integer
          minimum: 10
          maximum: 60
          description: Time-to-live in seconds
          example: 30
        branding:
          $ref: '#/components/schemas/BrandingInfo'
        signature:
          $ref: '#/components/schemas/Base64String'
    TokenCreateResponse:
      type: object
      required:
        - tokenId
        - expiresAt
        - callReference
      properties:
        tokenId:
          type: string
          pattern: '^TK-[A-Za-z0-9]{16}$'
          example: 'TK-abc123XYZ789def'
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
        callReference:
          $ref: '#/components/schemas/CallReference'
    TokenVerifyResponse:
      type: object
      required:
        - tokenId
        - originatingCP
        - callerID
        - calledID
        - callReference
        - verified
        - expiresAt
      properties:
        tokenId:
          type: string
        originatingCP:
          $ref: '#/components/schemas/RCPID'
        callerID:
          $ref: '#/components/schemas/PhoneNumber'
        calledID:
          $ref: '#/components/schemas/PhoneNumber'
        callReference:
          $ref: '#/components/schemas/CallReference'
        verified:
          type: boolean
        branding:
          $ref: '#/components/schemas/BrandingInfo'
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
    RoutingRequest:
      type: object
      required:
        - messageId
        - timestamp
        - requestingCP
        - callerID
        - destinationNumber
        - callReference
        - mediaCapabilities
        - publicKey
      properties:
        messageId:
          type: string
          format: uuid
        timestamp:
          $ref: '#/components/schemas/Timestamp'
        requestingCP:
          $ref: '#/components/schemas/RCPID'
        callerID:
          $ref: '#/components/schemas/PhoneNumber'
        destinationNumber:
          $ref: '#/components/schemas/PhoneNumber'
        callReference:
          $ref: '#/components/schemas/CallReference'
        mediaCapabilities:
          $ref: '#/components/schemas/MediaCapabilities'
        publicKey:
          $ref: '#/components/schemas/Base64String'
        branding:
          $ref: '#/components/schemas/BrandingInfo'
        signature:
          $ref: '#/components/schemas/Base64String'
...
    RoutingResponse:
      type: object
      required:
        - accepted
        - callReference
        - timestamp
      properties:
        accepted:
          type: boolean
        callReference:
          $ref: '#/components/schemas/CallReference'
        connectionDetails:
          $ref: '#/components/schemas/ConnectionDetails'
        agreedCapabilities:
          $ref: '#/components/schemas/MediaCapabilities'
        timestamp:
          $ref: '#/components/schemas/Timestamp'
        signature:
          $ref: '#/components/schemas/Base64String'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              enum:
                - call_not_found
                - invalid_signature
                - expired_token
                - invalid_token
                - capacity_exceeded
                - unsupported_codec
                - number_not_found
                - location_unavailable
                - unauthorized_psap
                - timeout
                - rate_limit_exceeded
                - invalid_request
                - internal_error
            message:
              type: string
            timestamp:
              $ref: '#/components/schemas/Timestamp'
            requestId:
              type: string
              format: uuid
  responses:
    Unauthorized:
      description: Unauthorized - Invalid signature or credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until retry allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  examples:
    VerificationRequestBasic:
      summary: Basic verification request
      value:
        messageId: '550e8400-e29b-41d4-a716-446655440000'
        timestamp: '2025-11-30T21:30:00.000Z'
        version: '1.0'
        requestingCP: 'CP1-UK-0002'
        callerID: '+441234567890'
        calledID: '+447700900123'
        callReference: 'abc-123-def-456'
        signature: 'base64-encoded-signature-here'
    VerificationResponseVerified:
      summary: Call verified successfully
      value:
        verified: true
        callReference: 'abc-123-def-456'
        callerName: 'John Smith'
        callerOrg: 'ACME Corp'
        callPurpose: 'Account verification'
        trustLevel: 'verified'
        branding:
          displayName: 'ACME Support'
          logo: 'https://cdn.acme.com/logo.png'
          backgroundColor: '#0066cc'
          textColor: '#ffffff'
        timestamp: '2025-11-30T21:30:00.050Z'
        signature: 'base64-signature'
    CallNotFound:
      summary: Call not found (potential fraud)
      value:
        error:
          code: 'call_not_found'
          message: 'No matching call record found'
          timestamp: '2025-11-30T21:30:00.050Z'
          requestId: '123e4567-e89b-12d3-a456-426614174000'
    RoutingRequestAudio:
      summary: Audio call routing request
      value:
        messageId: '123e4567-e89b-12d3-a456-426614174000'
        timestamp: '2025-11-30T21:30:00.000Z'
        requestingCP: 'CP1-UK-0001'
        callerID: '+441234567890'
        destinationNumber: '+447700900123'
        callReference: 'xyz-789'
        mediaCapabilities:
          codecs: ['opus', 'g722']
          encryption: ['srtp-aes256']
          video: false
        publicKey: 'base64-ed25519-key'
        signature: 'base64-signature'
    RoutingRequestVideo:
      summary: Video call routing request
      value:
        messageId: '123e4567-e89b-12d3-a456-426614174000'
        timestamp: '2025-11-30T21:30:00.000Z'
        requestingCP: 'CP1-UK-0001'
        callerID: '+441234567890'
        destinationNumber: '+447700900123'
        callReference: 'xyz-789'
        mediaCapabilities:
          codecs: ['opus', 'h264']
          encryption: ['srtp-aes256']
          video: true
          maxBandwidth: 2000000
        publicKey: 'base64-ed25519-key'
        signature: 'base64-signature'
...